Add initial implementation of Twitter-like application features
- Configure application properties for PostgreSQL database connection
- Implement DTOs for creating likes and tweets
- Create Like and Tweet entities with necessary relationships
- Develop Like and Tweet services with corresponding controllers
- Add global exception handling for user not found scenarios
- Set up security configuration to permit all requests

+++ src/main/java/com/example/twitter_challenge/Repository/BookmarkRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface BookmarkRepository extends JpaRepository<Bookmark, Integer> {
+public interface BookmarkRepository extends JpaRepository<Bookmark, Long> {
     Optional<Bookmark> findByTweetIdAndUserId(Long tweetId, Long userId);
+
 }
